// src/domain/models/AuthenticationRecommendedAction.ts
export enum AuthenticationRecommendedAction {
  NONE = 'NONE',
  RETRY_LATER = 'RETRY_LATER',
  VALIDATE_PROFILE = 'VALIDATE_PROFILE',
  REFRESH_SESSION = 'REFRESH_SESSION',
  IMPORT_PROFILE = 'IMPORT_PROFILE',
  EXPORT_PROFILE = 'EXPORT_PROFILE',
  EXPORT_NEW_PROFILE_AND_IMPORT = 'EXPORT_NEW_PROFILE_AND_IMPORT',
  REAUTHENTICATE = 'REAUTHENTICATE',
  RESTORE_PROFILE = 'RESTORE_PROFILE',
  CREATE_NEW_PROFILE = 'CREATE_NEW_PROFILE',
  CHECK_INFRASTRUCTURE = 'CHECK_INFRASTRUCTURE'
}